Fixed bugs and improved UI

See merge request fit3077-s1-2023/MA_Thursday4pm_Team7/project!13

+++ src/game/GameRuleRegulation/LegalMoves.java
@@ -29,6 +29,7 @@ public class LegalMoves {
         for(Position position : board.getPositionsCopy()) {
             if(position.getToken()==null) { // if no token
                 actionList.add(new PlaceAction(player,position));
+                position.setAllowed(true);
             }
         }
         return actionList;
@@ -42,6 +43,8 @@ public class LegalMoves {
                 for(Position endPos : startPos.getNeighboursCopy()) {
                     if(endPos.getToken()==null) {
                         actionList.add(new MoveAdjAction(player,startPos,endPos));
+                        startPos.setAllowed(true);
+                        endPos.setAllowed(true);
                     }
                 }
             }
@@ -57,6 +60,8 @@ public class LegalMoves {
                 for(Position endPos : board.getPositionsCopy()) {
                     if(endPos.getToken()==null) {
                         actionList.add(new FlyAction(player,startPos,endPos));
+                        startPos.setAllowed(true);
+                        endPos.setAllowed(true);
                     }
                 }
             }
@@ -68,8 +73,10 @@ public class LegalMoves {
         ArrayList<Action> actionList = new ArrayList<>();
         for(Position position : board.getPositionsCopy()) {
             if(position != null && position.getToken() != null
-                    && position.getToken().getPlayer() != player.getTeam()) {
+                    && position.getToken().getPlayer() != player.getTeam()
+                    && position.isMillExists()==false) {
                 actionList.add(new DeleteAction(player,position));
+                position.setAllowed(true);
             }
         }
         return actionList;
